-Solved Mail issue -Added count in patient APIs # Conflicts: #	dpdocter/src/main/java/com/dpdocter/elasticsearch/services/impl/ESPrescriptionServiceImpl.java #	dpdocter/src/main/java/com/dpdocter/request/Converter.java #	dpdocter/src/main/java/com/dpdocter/services/RecordsService.java #	dpdocter/src/main/java/com/dpdocter/services/impl/AppointmentServiceImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/BillingServiceImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/CertificateServicesImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/InventoryServiceImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/LocationServiceImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/PrescriptionServicesImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/RecordsServiceImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/SMSServicesImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/SignUpServiceImpl.java #	dpdocter/src/main/java/com/dpdocter/services/impl/UploadDataServicesimpl.java #	dpdocter/src/main/java/com/dpdocter/tokenstore/CustomPasswordEncoder.java #	dpdocter/src/main/java/common/util/web/AESCipherUtil.java

+++ dpdocter/src/main/java/com/dpdocter/services/impl/AdmitCardServiceImpl.java
@@ -603,9 +603,9 @@ public class AdmitCardServiceImpl implements AdmitCardService {
 			String body = mailBodyGenerator.generateEMREmailBody(mailResponse.getPatientName(),
 					mailResponse.getDoctorName(), mailResponse.getClinicName(), mailResponse.getClinicAddress(),
 					mailResponse.getMailRecordCreatedDate(), "Admit Card", "emrMailTemplate.vm");
-			mailService.sendEmail(emailAddress, mailResponse.getDoctorName() + " sent you Admit Card", body,
+			Boolean response = mailService.sendEmail(emailAddress, mailResponse.getDoctorName() + " sent you Admit Card", body,
 					mailResponse.getMailAttachment());
-			if (mailResponse.getMailAttachment() != null
+			if (response != null && mailResponse.getMailAttachment() != null
 					&& mailResponse.getMailAttachment().getFileSystemResource() != null)
 				if (mailResponse.getMailAttachment().getFileSystemResource().getFile().exists())
 					mailResponse.getMailAttachment().getFileSystemResource().getFile().delete();
@@ -688,9 +688,9 @@ public class AdmitCardServiceImpl implements AdmitCardService {
 			String body = mailBodyGenerator.generateEMREmailBody(mailResponse.getPatientName(),
 					mailResponse.getDoctorName(), mailResponse.getClinicName(), mailResponse.getClinicAddress(),
 					mailResponse.getMailRecordCreatedDate(), "Admit Card", "emrMailTemplate.vm");
-			mailService.sendEmail(emailAddress, mailResponse.getDoctorName() + " sent you Admit Card", body,
+			Boolean response = mailService.sendEmail(emailAddress, mailResponse.getDoctorName() + " sent you Admit Card", body,
 					mailResponse.getMailAttachment());
-			if (mailResponse.getMailAttachment() != null
+			if (response != null && mailResponse.getMailAttachment() != null
 					&& mailResponse.getMailAttachment().getFileSystemResource() != null)
 				if (mailResponse.getMailAttachment().getFileSystemResource().getFile().exists())
 					mailResponse.getMailAttachment().getFileSystemResource().getFile().delete();
